+++ test/wasmfs/wasmfs_readfile.c
@@ -6,21 +6,19 @@
  */
 
 #include <assert.h>
-#include <emscripten.h>
-#include <emscripten/wasmfs.h>
+#include <emscripten/em_asm.h>
 #include <errno.h>
 #include <fcntl.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <sys/stat.h>
 #include <unistd.h>
 
-#include "get_backend.h"
-
 int main() {
-  int err = wasmfs_create_directory("/root", 0777, get_backend());
+  int err = mkdir("root", 0777);
   assert(err == 0);
-  int fd = open("/root/test", O_CREAT | O_WRONLY, S_IRUSR | S_IWUSR);
+  int fd = open("root/test", O_CREAT | O_WRONLY, S_IRUSR | S_IWUSR);
   const char* msg = "Success\n";
 
   errno = 0;
@@ -30,25 +28,29 @@ int main() {
   assert(err == 0);
 
   EM_ASM({
-    var output = FS.readFile("/root/test");
+    var output = FS.readFile("root/test");
     out(UTF8ArrayToString(output));
     out("Length: " + output.byteLength);
-    var err = FS.unlink("/root/test");
+    var err = FS.unlink("root/test");
     out("FS.unlink: " + err);
   });
 
   EM_ASM({
     try {
-      var output = FS.readFile("/root/no-exist", {encoding : 'utf8'});
+      var output = FS.readFile("root/no-exist", {encoding : 'utf8'});
     } catch (err) {
+      return;
     }
+    assert(false);
   });
 
   EM_ASM({
     try {
       // Already unlinked above, file should not exist anymore
-      var output = FS.readFile("/root/test", {encoding : 'utf8'});
+      var output = FS.readFile("root/test", {encoding : 'utf8'});
     } catch (err) {
+      return;
     }
+    assert(false);
   });
 }